requirements.txt
fastapi==0.103.2
uvicorn==0.25.0
sqlalchemy==2.0.22
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.3
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.1.2
loguru==0.7.2
fastapi-cors==0.1.0
python-dotenv==1.0.1